:art: Longに変更

+++ backend/src/main/kotlin/com/uramnoil/serverist/graphql/Pagenation.kt
@@ -1,5 +1,5 @@
 package com.uramnoil.serverist.graphql
 
-data class PageRequest(val limit: Int, val offset: Int)
+data class PageRequest(val limit: Int, val offset: Long)
 
-data class PageInfo(val limit: Int, val offset: Int)
+data class PageInfo(val limit: Int, val offset: Long)
